Add files via upload
Scripts to pickup objects. Pickupable is empty, it is just an identifier so Unity knows what objects can be picked up. In other words, if a game object has the Pickupable script, we can pick it up. 

PickupObject is added to the RigidBodyFPSController and does all of the work related to picking up an object.